Always run fetch on repos in deps2git
When run with the --repos command, for each repo,
the first time deps2git is run, it clones the directory, then
crashes on the revision lookup.  The second time it is run, it
runs fetch on the directory, looks up the revision and moves on
to the next repo.

We fix this behavior by always running fetch, even on the initial checkout.

Review URL: https://chromiumcodereview.appspot.com/12225067

git-svn-id: http://src.chromium.org/svn/trunk/tools/deps2git@181098 4ff67af0-8c30-449e-8e8b-ad334ec8d88c