gitdeps: Only fetch repo if needed
Relanding earlier attempt to reduce fetches.

This does not work perfectly -- about 8 repos have DEPS pinned
to a revision which does not exist in the subfolder -- in this
case we cannot tell if we are missing the revision because of
a lost fetch, so we have to fetch. Once we use the same repos
as git_updater we can elimnate the fetches.

git-svn-id: http://src.chromium.org/svn/trunk/tools/deps2git@208952 4ff67af0-8c30-449e-8e8b-ad334ec8d88c